terminal corrections

+++ sql/omniva_int_terminal.sql
@@ -4,8 +4,8 @@ CREATE TABLE IF NOT EXISTS `_DB_PREFIX_omniva_int_terminal` (
     `city` varchar(100),
     `country_code` varchar(3),
     `address` varchar(255),
-    `x_cord` float(10),
-    `y_cord` float(10),
+    `x_cord` float,
+    `y_cord` float,
     `comment` varchar(255),
     `identifier` varchar(50),
     PRIMARY KEY (`id`)
